+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to Sample Commander will be documented in this file.
 
 ### Added
 * Files can now be selected by drawing a selection box around them by mouse (as in Windows Explorer, including the inversion behavior when holding the `Ctrl` key while drawing).
+* The vertical space between files can now be customized, e.g. for easier drawing of selection boxes. 
 
 ### Changed
 * removed "SimpleSelect" mode (selecting files by simple click without `Shift` or `Ctrl` pressed)